requirements.txt
streamlit>=1.28.0
pydantic>=2.0.0
requests>=2.31.0
folium>=0.14.0
streamlit-folium>=0.15.0
pytest>=7.4.0
python-dotenv>=1.0.0
openai>=1.0.0
pandas>=2.0.0
flask>=3.0.0
flask-cors>=4.0.0
